refactor(desktop): remove custom-electron-titlebar and simplify theme handling
- Remove entire titlebar folder with custom-electron-titlebar implementation
- Move getBackgroundColor utility to libs/utils.ts for reuse
- Add changeTheme method to DesktopApiSystem for centralized theme handling
- Remove legacy IPC-based theme update in favor of desktopApiProxy
- Clean up CSS by removing unused .cet-container and .cet-titlebar styles
- Fix code formatting in Menu.tsx and use useMemo for allItems
