--- apps/desktop/app/titlebar/index.ts
@@ -1,9 +1,0 @@
-/* ---------------------------------------------------------------------------------------------
- *  Copyright (c) AlexTorresDev. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *-------------------------------------------------------------------------------------------- */
-
-import { CustomTitlebar } from './titlebar';
-import { Color } from './base/common/color';
-
-export { CustomTitlebar, CustomTitlebar as Titlebar, Color as TitlebarColor };
