--- apps/desktop/app/titlebar/main/index.ts
@@ -1,9 +1,0 @@
-/* ---------------------------------------------------------------------------------------------
- *  Copyright (c) AlexTorresDev. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *-------------------------------------------------------------------------------------------- */
-
-import attachTitlebarToWindow from './attach-titlebar-to-window';
-import setupTitlebar from './setup-titlebar';
-
-export { setupTitlebar, attachTitlebarToWindow };
