+++ package.json
@@ -125,7 +125,6 @@
     "browserify-zlib": "^0.2.0",
     "colord": "^2.9.3",
     "crypto-browserify": "^3.12.0",
-    "custom-electron-titlebar": "4.2.8",
     "ejs-loader": "^0.5.0",
     "electron-taskbar-badge": "1.1.2",
     "ethersV6": "npm:ethers@^6.11.1",
